requirements.txt
aniso8601==9.0.1
appdirs==1.4.4
argcomplete==3.0.8
attrs==19.3.0
audioread==3.0.0
Automat==0.8.0
beautifulsoup4==4.12.2
blinker==1.6.2
blis==0.7.9
boto3==1.26.133
botocore==1.29.134
catalogue==2.0.8
certifi==2019.11.28
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.3
colorama==0.4.3
confection==0.0.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.7
cryptography==2.8
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.16
decorator==5.1.1
distro==1.4.0
en-core-web-md==3.5.0
entrypoints==0.3
filelock==3.12.0
Flask==2.3.2
flask-restx==1.1.0
fonttools==4.39.4
gdown==4.7.1
gTTS==2.3.2
h5py==3.8.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==6.6.0
importlib-resources==5.12.0
incremental==16.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kaldiio==2.18.0
keyring==18.0.1
kiwisolver==1.4.4
langcodes==3.3.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.2
librosa==0.10.0.post2
lit==16.0.3
llvmlite==0.40.0
macaroonbakery==1.3.1
MarkupSafe==2.1.2
matplotlib==3.7.1
mpmath==1.3.0
msgpack==1.0.5
munch==3.0.0
murmurhash==1.0.9
netifaces==0.10.4
networkx==3.1
numba==0.57.0
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.1.0
packaging==23.1
parallel-wavegan==0.5.5
pathy==0.10.1
pexpect==4.6.0
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
ply==3.11
pooch==1.6.0
preshed==3.0.8
protobuf==3.20.3
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
pycurl==7.43.0.2
pydantic==1.10.7
pydub==0.25.1
PyHamcrest==1.9.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
pyserial==3.4
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==5.3.1
requests==2.30.0
requests-unixsocket==0.2.0
s3transfer==0.6.1
scikit-learn==1.2.2
scipy==1.10.1
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SpeechRecognition==3.10.0
srsly==2.4.6
ssh-import-id==5.10
sympy==1.12
tensorboardX==2.6
thinc==8.1.10
threadpoolctl==3.1.0
tomlkit==0.11.8
torch==1.13.1
torchaudio==0.13.1
tqdm==4.65.0
triton==2.0.0
Twisted==18.9.0
typer==0.7.0
typing_extensions==4.5.0
ubuntu-advantage-tools==20.3
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
wasabi==1.1.1
Werkzeug==2.3.4
xkit==0.0.0
xmltodict==0.13.0
yq==3.2.2
zipp==3.15.0
zope.interface==4.7.1
